requirements.txt
Django==5.1
djangorestframework==3.15.2
jwt==1.3.1
channels[daphne]==4.1.0
